# src/api-specifications/api-spec.yml
openapi: '3.0.3'
info:
  title: DaaS-NodeJS
  version: '1.1'
  
  contact:
    name: API Support
    email: alessio.farfaglia@gmail.com
  
  license:
    name: MPL-2.0
    url: https://www.mozilla.org/en-US/MPL/2.0/
servers:
  - url: http://localhost:{port}/api
    description: Development server
    variables:
      port:
        default: '3000'
        description: default port for the development server
...
components:
  parameters:
    limit:
      name: limit
      in: query
      description: Numero massimo di oggetti da ritornare
      required: false
      schema:
        type: integer
        minimum: 1
        default: 20
      example: 20
    offset:
      name: offset
      in: query
      description: Offset per la paginazione
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      example: 0
  responses:  
    InternalError:
      description: Errore interno del server
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    DeviceModel:
      type: object
      properties:
        id:
          type: integer
          description: ID del modello del device
          nullable: false
        device_group_id:
          type: integer
          description: ID del gruppo del modello
          nullable: false
        description:
          type: string
          description: descrizione del modello
        serial:
          type: string
          description: seriale del modello
          maxLength: 20
...
    DeviceModelGroup:
      type: object
      properties:
        id:
          type: integer
          description: ID del gruppo del modello
          nullable: false
        title:
          type: string
          description: titolo del gruppo
          maxLength: 45
        description:
          type: string
          description: descrizione del gruppo
        link_image:
          type: string
          format: uri
          description: collegamento all'immagine del gruppo
          example: "https://placehold.co/600x400"
...
    PaginationData: 
      description: Meta dati sulla paginazione dei dati
      properties:
        count:
          type: integer
          description: numero di oggetti contenuti tra i dati della risposta
          minimum: 0
          example: 20
        offset:
          type: integer
          description: gli oggetti contenuti dei dati della risposta partono da questo offset
          minimum: 40
        limit:
          type: integer
          description: numero massimo di oggetti contenuti tra i dati della risposta
          minimum: 0
          example: 20
        total:
          description: numero totale di oggetti per questa query
          type: integer
          minimum: 0
          example: 67
        has_next:
          type: boolean
          description: flag che indica se ci sono altri oggetti da visualizzare
          example: true
        has_prev:
          type: boolean
          description: flag che indica se ci sono oggetti precedenti da visualizzare
          example: false
    Error:
      type: object
      properties:
        error_name:
          type: string
          description: Nome dell'errore
          example: "TypeError"
        message:
          type: string
          description: Messaggio di errore

#region Paths
paths:
...
  /configure:
    post:
      description: Carica la configurazione del nodo locale (per il din_local con id=1)
      tags:
        - configuration
      responses:
        '200':
          description: configurazione applicata.
        default:
          $ref: '#/components/responses/InternalError'
  /stop:
    post:
      description: Ferma il nodo daas locale
      tags:
        - lifecycle
      responses:
        '200':
          description: Nodo fermato.
        default:
          description: Impossibile fermare il nodo locale.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /start:
    post:
      description: Avvia il nodo daas locale
      tags:
        - lifecycle
      responses:
        '200':
          description: Nodo locale avviato.
        default:
          description: Impossibile avviare il nodo locale.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /send:
    post:
      description: Invia un messaggio tramite il nodo daas locale al DIN specificato
      tags:
        - commands
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                din:
                  type: integer
                  description: Il DIN a cui inviare il messaggio
                typeset:
                  type: integer
                  description: Il typeset del messaggio
                payload:
                  type: object
                  description: Il payload del messaggio
                  nullable: true
              required:
                - din
                - typeset
              example:
                din: 105
                typeset: 701
                payload: { "temperature": 25.5, "humidity": 50 }
      responses:
        '200':
          description: Messaggio inviato.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Messaggio inviato.
        default:
          $ref: '#/components/responses/InternalError'
  /status: 
    get:
      description: Ritorna lo status del nodo locale.
      tags:
        - lifecycle
      responses:
        '200':
          description: Infomazioni sullo stato del nodo.s
          content:
            application/json:
              schema:
                type: object
                properties:
                  lasttime:
                    type: number
                    description: time reference
                  hwver:
                    type: number
                    description: platform (hard-coded) when the node was contructed
                  linked:
                    type: number
                    description: channels counter (0=not linked)
                  sync:
                    type: number
                    description: synchronization status (0=not in sync)
                  lock:
                    type: number
                    description: required security policy
                  sklen:
                    type: number
                    description: security phrase length
                  skey:
                    type: string
                    description: security phrase (UTF-8)
                  form:
                    type: number
                    description: data formatting model
                  codec:
                    type: number
                    description: data encryption level
        default:
          $ref: '#/components/responses/InternalError'
...
  /dev/db_sync:
    post:
      description: >
        Valido solo in ambiente di sviluppo.
        
        
        Sincronizza il database locale con il modello di dati sequelize
      tags:
        - configuration
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                force:
                  type: boolean
                  description: Forza la sincronizzazione
              example:
                force: true
      responses:
        '200':
          description: Database sincronizzato.
        '403':
          description: Non autorizzato.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/InternalError'
#region Receivers
...
  /device_model_groups:
    get:
      description: Ritorna una pagina di tutti i gruppi di modelli di device.
      tags:
        - device model groups
      parameters:
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/limit'
        - name: q
          in: query
          description: >
            Filtro per nome del gruppo di modelli di device.
          schema:
            type: string
          example: "Gruppo 1"
      responses:
        '200':
          description: Pagina di gruppi di modelli di device e informazioni sulla paginazione
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeviceModelGroup'
                  pagination:
                    $ref: '#/components/schemas/PaginationData'
                  q:
                    type: string
                    description: Filtro applicato
                    nullable: true
        default:
          $ref: '#/components/responses/InternalError'

    post:
      description: Crea un nuovo gruppo di modelli di device.
      tags:
        - device model groups
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Nome del gruppo
                  maxLength: 45
                description:
                  type: string
                  description: Descrizione del gruppo
                  maxLength: 45
                link_image:
                  type: string
                  format: uri
                  description: Collegamento all'immagine del gruppo
              required:
                - name
              example:
                name: "Gruppo 1"
      responses:
        '200':
          description: Gruppo di modelli di device aggiunto con successo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceModelGroup'
        '400':
          description: Errore nei parametri della richiesta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/InternalError'
  /device_model_groups/{deviceModelGroupId}:
    parameters:
      - name: deviceModelGroupId
        in: path
        description: ID del gruppo di modelli di device
        required: true
        schema:
          type: integer
    get:
      description: Ritorna un gruppo di modelli di device specifico.
      tags:
        - device model groups      
      responses:
        '200':
          description: Gruppo di modelli di device trovato.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceModelGroup'
        '404':
          description: Gruppo di modelli di device non trovato.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/InternalError'
    
    put:
      description: Aggiorna un gruppo di modelli di device specifico.
      tags:
        - device model groups
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Nome del gruppo
                  maxLength: 45
                description:
                  type: string
                  description: Descrizione del gruppo
                  maxLength: 45
                link_image:
                  type: string
                  format: uri
                  description: Collegamento all'immagine del gruppo
              example:
                name: "Gruppo 1"
      responses:
        '200':
          description: Gruppo di modelli di device aggiornato con successo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Messaggio di successo
        '400':
          description: Errore nei parametri della richiesta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Gruppo di modelli di device non trovato.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/InternalError'

    delete:
      description: Elimina un gruppo di modelli di device specifico.
      tags:
        - device model groups
      responses:
        '200':
          description: Gruppo di modelli di device eliminato con successo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Messaggio di successo
        '404':
          description: Gruppo di modelli di device non trovato.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/InternalError'
  /device_model_groups/{deviceModelGroupId}/device_models:
    get:
      description: Ritorna i modelli di device associati al gruppo di modelli di device specificato
      tags:
        - device model groups
        - device models
      parameters:
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/limit'
        - name: q
          in: query
          description: >
            Filtro per nome del modello di device.
          schema:
            type: string
          example: "Modello 2"
      responses:
        '200':
          description: Pagina di modelli di device e informazioni sulla paginazione
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeviceModel'
                  pagination:
                    $ref: '#/components/schemas/PaginationData'
                  q:
                    type: string
                    description: Filtro applicato
                    nullable: true
        '404':
          description: Gruppo di modelli di device non trovato.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'        
        default:
          $ref: '#/components/responses/InternalError'

#endregion Device Model Groups




#region Programming
...